# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DOT Affiliates - FEARLESS WEEK 2.0 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerAffiliate
      tags: [auth]
      summary: Register a new affiliate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginAffiliate
      tags: [auth]
      summary: Login an affiliate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session affiliate
      responses:
        "200":
          description: Current affiliate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affiliate"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /affiliate/me:
    get:
      operationId: getAffiliateMe
      tags: [affiliate]
      summary: Get own affiliate profile and stats
      responses:
        "200":
          description: Profile with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateDashboard"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /affiliate/settings:
    put:
      operationId: updateAffiliateSettings
      tags: [affiliate]
      summary: Update affiliate settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affiliate"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /affiliate/password:
    put:
      operationId: changePassword
      tags: [affiliate]
      summary: Change affiliate password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    AffiliateRegistration:
      type: object
      required:
        - name
        - username
        - email
        - password
        - whatsappNumber
        - primaryPlatform
        - avgEngagement
        - hasPromotedBefore
        - whatsappGroupsReach
        - ticketsSellEstimate
        - estimatedReach
        - willingToPromote
        - whySelectYou
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        whatsappNumber:
          type: string
        phoneNumber:
          type: ["string", "null"]
        primaryPlatform:
          type: string
          enum: [instagram, tiktok, twitter, snapchat, whatsapp, facebook]
        avgEngagement:
          type: string
        hasPromotedBefore:
          type: boolean
        whatsappGroupsReach:
          type: string
        ticketsSellEstimate:
          type: string
          enum: ["0-50", "50-100", "100+"]
        estimatedReach:
          type: ["string", "null"]
        willingToPromote:
          type: boolean
        whySelectYou:
          type: string
    LoginInput:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or Username
        password:
          type: string
    AuthResponse:
      type: object
      required:
        - token
      properties:
        affiliate:
          $ref: "#/components/schemas/Affiliate"
        token:
          type: string
        role:
          type: string
          enum: [admin, affiliate]
    Affiliate:
      type: object
      required:
        - id
        - name
        - username
        - email
        - affiliateCode
        - status
        - primaryPlatform
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: string
        whatsappNumber:
          type: string
        phoneNumber:
          type: ["string", "null"]
        affiliateCode:
          type: string
        status:
          type: string
          enum: [pending, active, suspended]
        primaryPlatform:
          type: string
        avgEngagement:
          type: string
        hasPromotedBefore:
          type: boolean
        whatsappGroupsReach:
          type: string
        ticketsSellEstimate:
          type: string
        estimatedReach:
          type: ["string", "null"]
        willingToPromote:
          type: boolean
        whySelectYou:
          type: string
        createdAt:
          type: string
          format: date-time
    AffiliateDashboard:
      type: object
      required:
        - affiliate
        - stats
      properties:
        affiliate:
          $ref: "#/components/schemas/Affiliate"
        stats:
          $ref: "#/components/schemas/AffiliateStats"
    AffiliateStats:
      type: object
      required:
        - clicks
        - conversions
        - rank
        - affiliateLink
      properties:
        clicks:
          type: integer
        conversions:
          type: integer
        rank:
          type: ["integer", "null"]
        affiliateLink:
          type: string
...
    AffiliateSettingsUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        whatsappNumber:
          type: string
        phoneNumber:
          type: ["string", "null"]
        primaryPlatform:
          type: string
          enum: [instagram, tiktok, twitter, snapchat, whatsapp, facebook]
    PasswordChangeInput:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...