# lib/api-zod/src/generated/types/affiliate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { AffiliateStatus } from "./affiliateStatus";

export interface Affiliate {
  id: number;
  name: string;
  username: string;
  email: string;
  whatsappNumber?: string;
  /** @nullable */
  phoneNumber?: string | null;
  affiliateCode: string;
  status: AffiliateStatus;
  primaryPlatform: string;
  avgEngagement?: string;
  hasPromotedBefore?: boolean;
  whatsappGroupsReach?: string;
  ticketsSellEstimate?: string;
  /** @nullable */
  estimatedReach?: string | null;
  willingToPromote?: boolean;
  whySelectYou?: string;
  createdAt: Date;
}