# lib/api-zod/src/generated/types/affiliateRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { AffiliateRegistrationPrimaryPlatform } from "./affiliateRegistrationPrimaryPlatform";
import type { AffiliateRegistrationTicketsSellEstimate } from "./affiliateRegistrationTicketsSellEstimate";

export interface AffiliateRegistration {
  name: string;
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  whatsappNumber: string;
  /** @nullable */
  phoneNumber?: string | null;
  primaryPlatform: AffiliateRegistrationPrimaryPlatform;
  avgEngagement: string;
  hasPromotedBefore: boolean;
  whatsappGroupsReach: string;
  ticketsSellEstimate: AffiliateRegistrationTicketsSellEstimate;
  /** @nullable */
  estimatedReach: string | null;
  willingToPromote: boolean;
  whySelectYou: string;
}