# lib/api-zod/src/generated/types/affiliateSettingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { AffiliateSettingsUpdatePrimaryPlatform } from "./affiliateSettingsUpdatePrimaryPlatform";

export interface AffiliateSettingsUpdate {
  name?: string;
  username?: string;
  whatsappNumber?: string;
  /** @nullable */
  phoneNumber?: string | null;
  primaryPlatform?: AffiliateSettingsUpdatePrimaryPlatform;
}