# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { Affiliate } from "./affiliate";
import type { AuthResponseRole } from "./authResponseRole";

export interface AuthResponse {
  affiliate?: Affiliate;
  token: string;
  role?: AuthResponseRole;
}