# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOT Affiliates - FEARLESS WEEK 2.0 API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /** Email or Username */
  identifier: string;
  password: string;
}